--- perun-cli-python/cli/getPerunStatus.py
@@ -1,9 +1,0 @@
-from rich import print
-import cli
-
-
-def main() -> None:
-	""" prints perun status"""
-	rpc = cli.rpc
-	print('Status from', rpc.config.host)
-	print(rpc.utils.get_perun_status())
